+++ Server/src/controllers/Filters/filter.js
@@ -4,21 +4,45 @@ const filter = async (req, res) => {
   try {
     const { name, size, price } = req.query;
     const products = await Product.findAll();
+
     if (name) {
-      const namefiltered = products.filter((product) => product.name === name);
-      return res.status(200).json({ namefiltered });
+      const nameFiltered = products.filter((product) =>
+        product.name.toLowerCase().includes(name.toLowerCase())
+      );
+      if (nameFiltered.length > 0) {
+        return res.status(200).json({ nameFiltered });
+      } else {
+        return res
+          .status(404)
+          .json({ message: "No hay productos con ese Nombre" });
+      }
     }
 
     if (price) {
-      const pricefiltered = products.filter(
-        (product) => product.price === price
+      const numericPrice = parseFloat(price);
+      const priceFiltered = products.filter(
+        (product) => parseFloat(product.price) === numericPrice
       );
-      return res.status(200).json({ pricefiltered });
+      if (priceFiltered.length > 0) {
+        return res.status(200).json({ priceFiltered });
+      } else {
+        return res
+          .status(404)
+          .json({ message: "No hay productos con ese Precio" });
+      }
     }
 
-    if (size) {
-      const sizefiltered = products.filter((product) => product.size === size);
-      return res.status(200).json({ sizefiltered });
+    if (size && Array.isArray(size)) {
+      const sizeFiltered = products.filter((product) => {
+        product.size.some((productSize) => size.includes(productSize));
+      });
+      if (sizeFiltered.length > 0) {
+        return res.status(200).json({ sizeFiltered });
+      } else {
+        return res
+          .status(404)
+          .json({ message: "No hay productos con ese Talle" });
+      }
     }
   } catch (error) {
     return res.status(500).json({ error: error.message });
